FABIK/requirements.txt
Django
requests
google-auth
google-auth-oauthlib
python-dotenv
firebase-admin
whitenoise
dj-database-url
